backend/requirements.txt
fastapi
bcrypt
easyocr
google-genai
numpy
opencv-python-headless
pandas
pillow
python-dotenv
python-multipart
pymupdf
uvicorn
